added int64 scalar type

+++ graph/model/models.go
@@ -2,7 +2,7 @@ package model
 
 type Repository struct {
 	ID              string          `json:"id"`
-	RestAPIID       *int            `json:"rest_api_id"`
+	RestAPIID       *int64          `json:"rest_api_id"`
 	Name            *string         `json:"name"`
 	FullName        *string         `json:"fullName"`
 	CollaboratorsID *int            `json:"collaborators"`
